Initialised project and added gradle 7.5.1 (kotlin dsl).
