--- Java/Vehicle.java
@@ -1,5 +1,0 @@
-
-public interface Vehicle {
-
-  public String getType();
-}
