+++ Java/app/src/main/java/calculator/Car.java
@@ -1,0 +1,11 @@
+package calculator;
+
+import calculator.Vehicle;
+
+public class Car implements Vehicle
+{
+  @Override
+  public String getType() {
+    return "Car";
+  }
+}
