+++ Java/app/src/main/java/calculator/Vehicle.java
@@ -1,0 +1,6 @@
+package calculator;
+
+public interface Vehicle {
+
+  public String getType();
+}
