+++ Java/settings.gradle.kts
@@ -1,0 +1,2 @@
+rootProject.name = "toll-calculator-master"
+include("app")
